refactor store to use one reducer
