+++ client/src/main.jsx
@@ -7,13 +7,6 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import "bootstrap/dist/js/bootstrap.min.js";
 import "bootstrap-icons/font/bootstrap-icons.css";
 
-import {
-  ApolloClient,
-  InMemoryCache,
-  ApolloProvider,
-  createHttpLink,
-} from '@apollo/client';
-import { setContext } from '@apollo/client/link/context';
 
 import App from './App.jsx'
 import Error from "./pages/Error.jsx";
@@ -21,26 +14,10 @@ import Home from "./pages/Home.jsx";
 import Campaigns from "./pages/Campaigns.jsx"
 import Characters from "./pages/Characters.jsx"
 import Sessions from "./pages/Sessions.jsx"
+import Login from "./pages/Login.jsx";
+import StoreProvider from "./redux/GlobalState";
 
-
-const httpLink = createHttpLink({
-  uri: '/graphql'
-});
-
-const authLink = setContext((_, { headers }) => {
-  const token = auth.currentUser?.uid;
-  return {
-    headers: {
-      ...headers,
-      authorization: token ? `Bearer ${token}` : ''
-    }
-  }
-});
-
-const client = new ApolloClient({
-  link: authLink.concat(httpLink),
-  cache: new InMemoryCache()
-})
+import ROUTES from "./utils/routes.js";
 
 const router = createBrowserRouter([
   {
@@ -53,15 +30,19 @@ const router = createBrowserRouter([
         element: <Home />
       },
       {
-        path: "/campaigns",
+        path: ROUTES.CAMPAIGNS,
         element: <Campaigns />
       },
       {
-        path: "/characters",
+        path: ROUTES.CHARACTERS,
         element: <Characters />
       },
       {
-        path: "/sessions",
+        path: ROUTES.LOGIN,
+        element: <Login />
+      },
+      {
+        path: ROUTES.SESSIONS,
         element: <Sessions />
       },
     ]
@@ -70,8 +51,8 @@ const router = createBrowserRouter([
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <ApolloProvider client={client}>
+    <StoreProvider>
       <RouterProvider router={router} />
-    </ApolloProvider>
+    </StoreProvider>
   </React.StrictMode>,
 )
