+++ client/src/pages/Login.jsx
@@ -3,14 +3,17 @@ import { useEffect, useState } from "react";
 import { Modal } from "bootstrap/dist/js/bootstrap.min.js";
 import { signInWithEmailAndPassword, createUserWithEmailAndPassword, sendPasswordResetEmail } from "firebase/auth";
 import { auth } from "../../../firebase.js"
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { USER_ACTIONS } from "../redux/reducers/userReducer.js";
 import { useLazyQuery } from "@apollo/client";
 import { GET_ME } from "../utils/queries.js";
+import { useNavigate } from "react-router-dom";
+import ROUTES from "../utils/routes.js";
 
 export default function Login() {
 
   const dispatch = useDispatch();
+  const navigate = useNavigate();
   const [getMe, { loading, data: userData }] = useLazyQuery(GET_ME);
 
   const [showSignup, setSignup] = useState(false);
@@ -24,7 +27,6 @@ export default function Login() {
   const [signupPassword, setSignupPassword] = useState('');
 
   useEffect(() => {
-    console.log("userData:", userData);
     if (userData) {
       console.log("userData.getMe:", userData.getMe);
       const userInfo = userData.getMe;
@@ -32,6 +34,8 @@ export default function Login() {
         type: USER_ACTIONS.LOGIN,
         user: userInfo
       });
+
+      // navigate(ROUTES.CHARACTERS);
     }
   }, [userData])
