+++ client/src/redux/GlobalState.jsx
@@ -1,7 +1,6 @@
 import { Provider } from 'react-redux' ;
 import store from './store'
 
-
 export default function StoreProvider(props) {
   return <Provider store={store} {...props} />;
 }
