+++ client/src/redux/reducer.js
@@ -1,4 +1,8 @@
-import { initialState } from "../state.js";
+export const USER_ACTIONS = {
+  LOGIN: "LOGIN",
+  SIGNUP: "SIGNUP",
+  LOGOUT: "LOGOUT",
+}
 
 export const CHARACTER_ACTIONS = {
   CREATE: "CREATE",
@@ -8,39 +12,66 @@ export const CHARACTER_ACTIONS = {
   BACKUP: "BACKUP"
 }
 
-export const characterReducer = (state = initialState, action) => {
+const initialState = {
+  user: null,
+  characters: null,
+}
+
+export const reducer = (state = initialState, action) => {
   switch (action.type) {
+    case USER_ACTIONS.LOGIN:
+      console.log("@user login");
+      return {
+        ...state,
+        user: action.user
+      }
+
+    case USER_ACTIONS.SIGNUP:
+      console.log("@user signup");
+      return {
+        ...state,
+        user: action.user
+      }
+
+    case USER_ACTIONS.LOGOUT:
+      console.log("@user logout");
+      return {
+        ...state,
+        user: null
+      }
+
+    /********** CHARACTER **********/
     case CHARACTER_ACTIONS.CREATE:
       console.log("@character create");
-      
+
       return {
         ...state,
       }
 
     case CHARACTER_ACTIONS.GET:
       console.log("@character get");
-      
+
       return {
         ...state,
       }
 
     case CHARACTER_ACTIONS.EDIT:
       console.log("@character edit");
-      
+
       return {
         ...state,
       }
 
     case CHARACTER_ACTIONS.DELETE:
       console.log("@character delete");
-      
+
       return {
         ...state,
       }
 
     case CHARACTER_ACTIONS.BACKUP:
       console.log("@character backup");
-      
+
       return {
         ...state,
       }
