--- client/src/redux/state.js
@@ -1,4 +1,0 @@
-export const initialState = {
-  user: null, // 
-  characters: null,
-}
