+++ client/src/utils/routes.js
@@ -2,6 +2,7 @@ const ROUTES = {
   CAMPAIGNS: "/campaigns",
   CHARACTERS: "/characters",
   HOME: "/",
+  LOGIN: "/login",
   SESSIONS: "/sessions"
 }
